Fix bug in README Makevars code snippet
The call to `cat(..., append = FALSE)` fails if the existing contents of `Makevars` are being read on the fly using `readLines(M)`. I'm guessing this is because the desired contents have already been overwritten by the preceding strings passed to `cat()`, although this doesn't quite make sense. In any case, the revised version with the contents of `M` assigned to a variable first works as intended.