Modify priors on initial age structure states to avoid divergences
For unclear reasons, the "folded" prior that allowed a single simplex array q_MR_init to be used for both the full [maiden | repeat] age structure in year 1 and the maiden-only age structure in years 2:max_age resulted in huge numbers of divergences. Changing to two separate arrays of simplices, the first of length N_MRage and the second of length N_age, is less elegant but makes the excess divergences disappear.