Allow user-specified age-selectivity in harvest and broodstock collection
As proposed in Issue #11, this approach uses arguments F_ages and B_ages to specify which adult ages are fully (non)selected, replacing the legacy code that hardwired the youngest age "jacks" as nonselected, albeit inconsistently across models. The default is all ages selected, so when this feature is merged into master it will require adding the new arguments to salmonIPM() calls in projects where jacks are nonselected.