Enable user-specified priors in IPM_SS_np
First test of user-specified priors on selected hyperparameters:
* alpha
* Rmax
* mu_p
* mu_SS (if iteroparous)
* tau

Tested and works, but implementation in stan_data() is clunky. Will need to functionalize the code chunk that transforms the priors argument into Stan-readable form and assigns defaults if necessary; otherwise it will get too cumbersome when other models are added. But should this functionality be folded into a prior_summary() function for the eventual salmonIPMfit class, as in rstanarm?