Simplify Ricker RRS function (see Issue #56)
It turns out the "simpler" formulation that had been (inacccurately) presented in the Details section of SR() is in fact the version to use. It does have symmetric density-dependent effects by W and H spawners, and it does not suffer from the pathologies documented in #56.

I am currently unable to compile salmonIPM on my machine, so committing these changes on a provisional basis for now.