Workaround for bug in nested max(int, int) calls that was causing compilation error in R 4.3.2
Compilation has been failing with an inscrutable error message on my machine since upgrading to R 4.3.2 (with rstan 2.32.5 and StanHeaders 2.32.5). Tried compiling a standalone version of IPM_SS_pp.stan using stan_model() and this also failed. By incrementally removing code, I isolated the issue to statements of the form

max(max(int, int), int);

That is, nested integer-max function calls. These were found in the definitions of N_orphan_age. I replaced them with equivalent use of fabs() and to_int(), and now the package compiles normally.

Clearly this is a bug. I will try to see if others can reproduce it, and if so, report it on Stan Discourse.