Various (mostly) S3-related changes to pass `devtools::check()`
The most important lessons are:

* Import all packages whose functions (S3 generics or otherwise) are used. Then @importFrom specific functions. This does not *attach* the package when salmonIPM is loaded, and unless @export-ed its functions are not available to the user.

* The arguments to a S3 method must include all of the named args from the generic (including ...). Additional args are OK.

The remaining WARNING says that all functions in draws.R after as_draws(), i.e. all as_draws_*(), as well as nsamples(), are undocumented. But these are associated with the corresponding main topic @name using @rdname, and they do show up in the doc, so ... ?