Expand anonymous arguments in saved call
Previously, if args such as priors were assigned to an object rather than defined inline, the saved call would not contain their values. In the case of priors, this meant that prior_summary() was looking for user-modified priors in object$call$priors and not finding them. Fixed by eval()ing all elements of the call, except data sets, in the parent environment.